% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_species.common.R
\name{classify_species.common}
\alias{classify_species.common}
\title{Classify Species in latin names based on common names.}
\usage{
classify_species.common(data, species_col = "Species")
}
\arguments{
\item{data}{A data frame with species names you want to make latin names}

\item{species_col}{The names column you want to base your classification on}
}
\description{
Classify Species in latin names based on common names.
}
\author{
Jonathan Dain \email{jonathan.dain001@umb.edu}
}
