% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_species.custom.R
\name{classify_species.custom}
\alias{classify_species.custom}
\title{Title}
\usage{
classify_species.custom(data, species_col = "Species", lookup_table)
}
\arguments{
\item{data}{A data grame with species names you want to convert to latin names.}

\item{species_col}{The common names column you want to base your classification on.}

\item{lookup_table}{A custom lookup table that you either created by hand or obtained from \code{update_lookup()}.}
}
\value{
Returns a dataframe with the classifications.
}
\description{
Title
}
