% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_species.scientific.R
\name{classify_species.scientific}
\alias{classify_species.scientific}
\title{Classify Species in common names based on latin names.}
\usage{
classify_species.scientific(data, species_col = "scientific.name")
}
\arguments{
\item{data}{A data frame with species names you want to make common names}

\item{species_col}{The names column you want to base your classification on}
}
\description{
Classify Species in common names based on latin names.
}
\author{
Jonathan Dain \email{jonathan.dain001@umb.edu}
}
