% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_species.type.R
\name{classify_species.type}
\alias{classify_species.type}
\title{Classify Species in functional groups (types) based on scientific names:}
\usage{
classify_species.type(data, species_col = "Species")
}
\arguments{
\item{data}{A data frame with species names you want to add the type too}

\item{species_col}{The names column you want to base your classification on}
}
\description{
Classify Species in functional groups (types) based on scientific names:
}
\author{
Jonathan Dain \email{jonathan.dain001@umb.edu}
}
