% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/days_to_run.R
\name{days_to_run}
\alias{days_to_run}
\title{BEAST days to run helper}
\usage{
days_to_run(total_steps, hours_per_million)
}
\arguments{
\item{total_steps}{The total number of steps to be performed by the MCMC chain. This is specified inside of the BEAST.xml file under MCMC chain.}

\item{hours_per_million}{This is the estimate "hours per million steps" value that is given in the print to screen output during a BEAST run.}
}
\value{
Returns an estimate number of days for a BEAST run to finish.
}
\description{
BEAST days to run helper
}
