% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_lookup_helper.R
\name{update_lookup}
\alias{update_lookup}
\title{Update Lookup Helper}
\usage{
update_lookup(
  common.name = "string",
  type = "string",
  scientific.name = "string",
  english.name.ebird = "string",
  order = "string",
  family,
  domestic = "string",
  lookup_table = NULL
)
}
\arguments{
\item{common.name}{an alternative common name, usually a misspelling of a traditional common name.}

\item{type}{the management type of the species i.e. (other,songbird, seabird, dabbling duck, raptor/owl/falcon, goose/swan, shorebird, gull, seaduck, chicken, domestic poultry, mammal, NA)}

\item{scientific.name}{the scientific name for this species. It is recommended to use https://birdsoftheworld.org for the most up to date taxonomy.}

\item{english.name.ebird}{the english name from https://birdsoftheworld.org for this species}

\item{order}{the https://birdsoftheworld.org taxonomic order for this species}

\item{family}{the https://birdsoftheworld.org taxonomic family for this species.}

\item{domestic}{whether this is a domestic or wild species.}

\item{lookup_table}{a user defined lookup table to allow for further editing. The default is set to NULL and utilizes the lookup table provided in the package.}
}
\value{
Returns the updated lookup table with the new row added. It is highly recommended that the user save this as a seperate lookup file to be used with \code{classify_species.custom()} in the future.
}
\description{
Update Lookup Helper
}
\author{
Jonathan Dain,
}
